# apps/server/internal/handler/notification.go
package handler

import (
	"context"
	"time"

	"github.com/gin-gonic/gin"

	"github.com/dingit-me/server/internal/model"
	"github.com/dingit-me/server/internal/pkg/pagination"
	"github.com/dingit-me/server/internal/pkg/response"
	"github.com/dingit-me/server/internal/service"
	"github.com/dingit-me/server/internal/ws"
)

// notificationStore is the subset of *service.Store that
// NotificationHandler actually touches. Defined as an unexported
// interface so unit tests can supply a lightweight in-memory mock
// without standing up a real PostgreSQL pool. Go structural typing
// means *service.Store automatically satisfies this interface, so
// main.go passes one in unchanged.
type notificationStore interface {
	Add(ctx context.Context, n *model.Notification) (*model.Notification, error)
	Get(ctx context.Context, id string) (*model.Notification, error)
	List(ctx context.Context, status *model.NotificationStatus, priority *model.NotificationPriority, limit, offset int) ([]model.Notification, error)
	Count(ctx context.Context, status *model.NotificationStatus, priority *model.NotificationPriority) (int, error)
	UpdateStatus(ctx context.Context, id string, status model.NotificationStatus, actionedValue *string) (*model.Notification, error)
	Delete(ctx context.Context, id string) (bool, error)
}

type NotificationHandler struct {
	store       notificationStore
	hub         *ws.Hub
	callbackSvc *service.CallbackService
	pushRouter  *service.PushRouter
}

func NewNotificationHandler(store notificationStore, hub *ws.Hub, callbackSvc *service.CallbackService, pushRouter *service.PushRouter) *NotificationHandler {
	return &NotificationHandler{store: store, hub: hub, callbackSvc: callbackSvc, pushRouter: pushRouter}
}

// detachedPush runs fn in a background goroutine with a context that
// survives the request completing or the client disconnecting, bounded
// by its own 30s timeout. fn receives a freshly-computed count of
// pending notifications so push-router calls can sync iOS badge numbers
// without each caller re-querying the store.
//
// context.WithoutCancel (Go 1.21+) strips the request's cancel signal
// while preserving request-scoped values (trace IDs, etc).
func (h *NotificationHandler) detachedPush(reqCtx context.Context, fn func(ctx context.Context, pendingCount int)) {
	bgCtx := context.WithoutCancel(reqCtx)
	go func() {
		ctx, cancel := context.WithTimeout(bgCtx, 30*time.Second)
		defer cancel()
		pending := model.StatusPending
		count, _ := h.store.Count(ctx, &pending, nil)
		fn(ctx, count)
	}()
}

// createResponse is the data field returned by POST /api/notifications.
type createResponse struct {
	ID        string `json:"id"        example:"550e8400-e29b-41d4-a716-446655440000"`
	Status    string `json:"status"    example:"pending"`
	Timestamp string `json:"timestamp" example:"2026-04-12T10:30:00Z"`
}

type createRequest struct {
	Title       string                     `json:"title" binding:"required"`
	Body        string                     `json:"body" binding:"required"`
	Source      string                     `json:"source"`
	Priority    string                     `json:"priority"`
	Icon        *string                    `json:"icon"`
	Actions     []model.NotificationAction `json:"actions"`
	CallbackURL *string                    `json:"callback_url"`
	Metadata    map[string]any             `json:"metadata"`
	TTL         *int                       `json:"ttl"`
	ExpiresAt   *string                    `json:"expires_at"`
}

// Create godoc
//
//	@Summary		Create a notification
//	@Description	Send a new notification to all connected clients via WebSocket and push.
//	@Tags			notifications
//	@Accept			json
//	@Produce		json
//	@Param			body	body		createRequest								true	"Notification payload"
//	@Success		201		{object}	response.Response{data=createResponse}		"Notification created"
//	@Failure		400		{object}	response.Response							"Validation error"
//	@Failure		401		{object}	response.Response							"Unauthorized"
//	@Failure		500		{object}	response.Response							"Internal error"
//	@Security		BearerAuth
//	@Router			/api/notifications [post]
func (h *NotificationHandler) Create(c *gin.Context) {
	var req createRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, response.CodeBadRequest, "title and body are required")
		return
	}

	// Reject unsafe callback URLs at ingress so we never persist a row
	// that points at an internal address. Fail-closed: if validation
	// can't resolve the DNS, we return BadRequest rather than storing it.
	if req.CallbackURL != nil && *req.CallbackURL != "" {
		if err := h.callbackSvc.Validate(*req.CallbackURL); err != nil {
			response.BadRequest(c, response.CodeBadRequest, "invalid callback_url: "+err.Error())
			return
		}
	}

	source := req.Source
	if source == "" {
		source = "unknown"
	}

	priority := model.NotificationPriority(req.Priority)
	if req.Priority == "" {
		priority = model.PriorityNormal
	} else if !isValidPriority(priority) {
		response.BadRequest(c, response.CodeInvalidPriority, "Invalid priority value. Allowed: urgent, high, normal, low")
		return
	}

	var expiresAt *time.Time
	if req.TTL != nil {
		if *req.TTL < 0 {
			response.BadRequest(c, response.CodeBadRequest, "TTL must be a positive number of seconds")
			return
		}
		if *req.TTL > 0 {
			t := time.Now().UTC().Add(time.Duration(*req.TTL) * time.Second)
			expiresAt = &t
		}
	} else if req.ExpiresAt != nil && *req.ExpiresAt != "" {
		t, err := time.Parse(time.RFC3339, *req.ExpiresAt)
		if err != nil {
			response.BadRequest(c, response.CodeBadRequest, "Invalid expires_at format, use RFC3339")
			return
		}
		if t.Before(time.Now()) {
			response.BadRequest(c, response.CodeBadRequest, "expires_at must be in the future")
			return
		}
		expiresAt = &t
	}

	n := &model.Notification{
		Title:       req.Title,
		Body:        req.Body,
		Source:      source,
		Priority:    priority,
		Icon:        req.Icon,
		Actions:     req.Actions,
		CallbackURL: req.CallbackURL,
		Metadata:    req.Metadata,
		ExpiresAt:   expiresAt,
	}

	created, err := h.store.Add(c.Request.Context(), n)
	if err != nil {
		response.InternalError(c, "Failed to create notification")
		return
	}

	h.hub.Broadcast(model.NewNotificationNewMsg(created))

	h.detachedPush(c.Request.Context(), func(ctx context.Context, count int) {
		h.pushRouter.SendToAll(ctx, created, count)
	})

	response.Created(c, gin.H{
		"id":        created.ID,
		"status":    created.Status,
		"timestamp": created.Timestamp,
	})
}

// List godoc
//
//	@Summary		List notifications
//	@Description	Retrieve a paginated list of notifications, optionally filtered by status and priority.
//	@Tags			notifications
//	@Produce		json
//	@Param			page		query		int		false	"Page number"			default(1)
//	@Param			page_size	query		int		false	"Items per page"		default(20)
//	@Param			status		query		string	false	"Filter by status"		Enums(pending,actioned,dismissed,expired)
//	@Param			priority	query		string	false	"Filter by priority"	Enums(urgent,high,normal,low)
//	@Success		200			{object}	response.Response{data=pagination.Result[model.Notification]}	"Paginated list"
//	@Failure		400			{object}	response.Response												"Invalid filter"
//	@Failure		401			{object}	response.Response												"Unauthorized"
//	@Failure		500			{object}	response.Response												"Internal error"
//	@Security		BearerAuth
//	@Router			/api/notifications [get]
func (h *NotificationHandler) List(c *gin.Context) {
	ctx := c.Request.Context()

	var p pagination.Params
	_ = c.ShouldBindQuery(&p)
	page, pageSize, offset := p.Normalize()

	var status *model.NotificationStatus
	if s := c.Query("status"); s != "" {
		st := model.NotificationStatus(s)
		if !isValidStatus(st) {
			response.BadRequest(c, response.CodeInvalidStatus, "Invalid status filter")
			return
		}
		status = &st
	}

	var priority *model.NotificationPriority
	if p := c.Query("priority"); p != "" {
		pr := model.NotificationPriority(p)
		if !isValidPriority(pr) {
			response.BadRequest(c, response.CodeInvalidPriority, "Invalid priority filter")
			return
		}
		priority = &pr
	}

	notifications, err := h.store.List(ctx, status, priority, pageSize, offset)
	if err != nil {
		response.InternalError(c, "Failed to list notifications")
		return
	}

	total, err := h.store.Count(ctx, status, priority)
	if err != nil {
		response.InternalError(c, "Failed to count notifications")
		return
	}

	response.Success(c, pagination.NewResult(notifications, int64(total), page, pageSize))
}

// GetByID godoc
//
//	@Summary		Get a notification
//	@Description	Retrieve a single notification by its ID.
//	@Tags			notifications
//	@Produce		json
//	@Param			id	path		string	true	"Notification ID"
//	@Success		200	{object}	response.Response{data=model.Notification}	"Notification found"
//	@Failure		401	{object}	response.Response							"Unauthorized"
//	@Failure		404	{object}	response.Response							"Not found"
//	@Failure		500	{object}	response.Response							"Internal error"
//	@Security		BearerAuth
//	@Router			/api/notifications/{id} [get]
func (h *NotificationHandler) GetByID(c *gin.Context) {
	id := c.Param("id")

	n, err := h.store.Get(c.Request.Context(), id)
	if err != nil {
		response.InternalError(c, "Failed to get notification")
		return
	}
	if n == nil {
		response.NotFound(c, "Notification not found")
		return
	}

	response.Success(c, n)
}

type updateRequest struct {
	Status        *model.NotificationStatus `json:"status"`
	ActionedValue *string                   `json:"actioned_value"`
}

// Update godoc
//
//	@Summary		Update notification status
//	@Description	Update the status of a notification (e.g. mark as actioned or dismissed).
//	@Tags			notifications
//	@Accept			json
//	@Produce		json
//	@Param			id		path		string										true	"Notification ID"
//	@Param			body	body		updateRequest								true	"Status update"
//	@Success		200		{object}	response.Response{data=model.Notification}	"Updated notification"
//	@Failure		400		{object}	response.Response							"Validation error"
//	@Failure		401		{object}	response.Response							"Unauthorized"
//	@Failure		404		{object}	response.Response							"Not found"
//	@Failure		500		{object}	response.Response							"Internal error"
//	@Security		BearerAuth
//	@Router			/api/notifications/{id} [patch]
func (h *NotificationHandler) Update(c *gin.Context) {
	id := c.Param("id")
	ctx := c.Request.Context()

	var req updateRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		response.BadRequest(c, response.CodeBadRequest, "Invalid request body")
		return
	}

	if req.Status == nil {
		response.BadRequest(c, response.CodeBadRequest, "status is required")
		return
	}

	if !isValidStatus(*req.Status) {
		response.BadRequest(c, response.CodeInvalidStatus, "Invalid status value")
		return
	}

	updated, err := h.store.UpdateStatus(ctx, id, *req.Status, req.ActionedValue)
	if err != nil {
		response.InternalError(c, "Failed to update notification")
		return
	}
	if updated == nil {
		response.NotFound(c, "Notification not found")
		return
	}

	h.hub.Broadcast(model.NewNotificationUpdatedMsg(updated))

	// Sync badge across all iOS devices when pending count changes.
	// Any status change away from "pending" reduces the count.
	if *req.Status == model.StatusDismissed || *req.Status == model.StatusActioned || *req.Status == model.StatusExpired {
		h.detachedPush(ctx, func(ctx context.Context, count int) {
			h.pushRouter.UpdateBadge(ctx, count)
		})
	}

	if *req.Status == model.StatusActioned && updated.CallbackURL != nil && req.ActionedValue != nil {
		h.callbackSvc.Deliver(updated, &model.ActionResponse{
			NotificationID: id,
			Action:         *req.ActionedValue,
		})
	}

	response.Success(c, updated)
}

// Delete godoc
//
//	@Summary		Delete a notification
//	@Description	Permanently remove a notification by its ID.
//	@Tags			notifications
//	@Produce		json
//	@Param			id	path	string	true	"Notification ID"
//	@Success		204	"No content"
//	@Failure		401	{object}	response.Response	"Unauthorized"
//	@Failure		404	{object}	response.Response	"Not found"
//	@Failure		500	{object}	response.Response	"Internal error"
//	@Security		BearerAuth
//	@Router			/api/notifications/{id} [delete]
func (h *NotificationHandler) Delete(c *gin.Context) {
	id := c.Param("id")

	deleted, err := h.store.Delete(c.Request.Context(), id)
	if err != nil {
		response.InternalError(c, "Failed to delete notification")
		return
	}
	if !deleted {
		response.NotFound(c, "Notification not found")
		return
	}

	h.hub.Broadcast(model.NewNotificationDeletedMsg(id))

	response.NoContent(c)
}

func isValidStatus(s model.NotificationStatus) bool {
	switch s {
	case model.StatusPending, model.StatusActioned, model.StatusDismissed, model.StatusExpired:
		return true
	}
	return false
}

func isValidPriority(p model.NotificationPriority) bool {
	switch p {
	case model.PriorityUrgent, model.PriorityHigh, model.PriorityNormal, model.PriorityLow:
		return true
	}
	return false
}